# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce API",
        "description": "API completa para o sistema de E-commerce Fullstack.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Listar produtos",
                "description": "Retorna todos os produtos cadastrados",
                "operationId": "a4aaa4f5e98e7123e7803e391a6e4e2d",
                "responses": {
                    "200": {
                        "description": "Lista de produtos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "abc123"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Notebook"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Notebook gamer"
                                            },
                                            "price": {
                                                "type": "number",
                                                "example": 5999.9
                                            },
                                            "stock": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "category_id": {
                                                "type": "string",
                                                "example": "cat123"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Criar produto",
                "operationId": "3f003bfc0928d1f8bf1867f4314a14c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "stock",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Notebook"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Notebook gamer"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 5999.9
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "category_id": {
                                        "type": "string",
                                        "example": "cat123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Buscar produto por ID",
                "operationId": "f33541f61b52d7d9e18530c19f694590",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abc123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto encontrado"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Atualizar produto",
                "operationId": "f45bfefc8566fe344a21cb1720b699b3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abc123"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Notebook atualizado"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição atualizada"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 4999.9
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Remover produto",
                "operationId": "962165106847db5ed0923292384cfb07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abc123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto removido"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Listar usuários",
                "operationId": "83119c699fe97cfb2fb004cebc287051",
                "responses": {
                    "200": {
                        "description": "Lista de usuários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "9b2c3e1a-8f72-4b0e-9c7e-5d8a9f0c3a21"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Matheus"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "matheus@email.com"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2026-03-25T12:00:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Buscar usuário por ID",
                "operationId": "0ec22789399cf07e1f632924a4ca2815",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "9b2c3e1a-8f72-4b0e-9c7e-5d8a9f0c3a21"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "9b2c3e1a-8f72-4b0e-9c7e-5d8a9f0c3a21"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Matheus"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "matheus@email.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2026-03-25T12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}